// app/api/auth/register/registerSchema.ts
import { z } from "zod";

export const registerSchema = z.object({
    name: z.string().min(3, "Name must be at least 3 characters long"),
    email: z.string(),
    password: z.string().min(6, "Password must be at least 6 characters long"),
})

export const registerResponse = z.object({
    success: z.boolean(),
    message: z.string(),
    data: registerSchema.omit({
        password: true
    }).optional()
})

export type RegisterResponseSchema = z.infer<typeof registerResponse>